third_party/WebKit/LayoutTests/animations/animationworklet/cancel-non-accelerated-property-expected.txt

CONSOLE WARNING: The animation cannot be accelerated. Reason: Accelerated keyframe value could not be computed
CONSOLE MESSAGE: line 49: background-color for the first target is: rgb(0, 255, 0)
